ADJUST - test adds {time} obj
